requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
Pillow==7.2.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqtgraph==0.11.0
pyrtlsdr==0.2.92
pyserial==3.4
python-dateutil==2.8.1
requests==2.24.0
scipy==1.5.2
six==1.15.0
urllib3==1.25.10
